A lot of small styling additions.
Added the navigation logo and navigation call-to-action button in the navbar. Gave the hero text a reduced maximum width, to keep it more narrow and focused. Gave the service cards some icons. Improved the size and positions of the temporary floating shapes in the About section. Gave the contact cards an envelope and phone icon, and bold header font. Whipped up a quick, temporary item list for the footer.